refactor: client input validation at server side
