+++ controllers/user-controller.js
@@ -1,30 +1,16 @@
 const bcrypt = require('bcryptjs')
 const dayjs = require('dayjs')
 const jwt = require('jsonwebtoken')
-const validator = require('validator')
+
 const helpers = require('../_helpers')
 const { User, Reply, Tweet, Like, Followship, sequelize } = require('../models')
 const { imgurFileHandler } = require('../helpers/file-helper')
 
 const userController = {
+
   signUp: async (req, res, next) => {
     try {
-      const { account, name, email, password, checkPassword } = req.body
-      if (!account.trim() || !name.trim() || !email.trim() || !password.trim() || !checkPassword.trim()) {
-        return res.status(400).json({ status: 'error', message: 'All field are required!' })
-      }
-      if (!validator.isEmail(email)) {
-        return res.status(422).json({ status: 'error', message: 'Email input is invalid!' })
-      }
-      if (password !== checkPassword) {
-        return res.status(422).json({ status: 'error', message: 'Password and confirmPassword do not match.' })
-      }
-      if (name.length > 50) {
-        return res.status(422).json({ status: 'error', message: 'Name field has max length of 50 characters.' })
-      }
-      const [userAccount, userEmail] = await Promise.all([User.findOne({ where: { account } }), User.findOne({ where: { email } })])
-      if (userAccount) return res.status(422).json({ status: 'error', message: 'Account already exists!' })
-      if (userEmail) return res.status(422).json({ status: 'error', message: 'Email already exists!' })
+      const { account, name, email, password } = req.body
 
       const hash = await bcrypt.hash(password, 10)
       let userRegistered = await User.create({ account, name, email, password: hash })
@@ -39,6 +25,7 @@ const userController = {
       return res.status(200).json({ status: 'success', data: { token, user: userRegistered } })
     } catch (err) { next(err) }
   },
+
   signIn: (req, res, next) => {
     try {
       const loginUser = helpers.getUser(req).toJSON()
@@ -53,6 +40,7 @@ const userController = {
       return res.status(200).json({ status: 'success', data: { token, user: loginUser } })
     } catch (err) { next(err) }
   },
+
   getUserProfile: async (req, res, next) => {
     try {
       const reqId = Number(req.params.id)
@@ -72,30 +60,37 @@ const userController = {
       return res.status(200).json(user)
     } catch (err) { next(err) }
   },
+
   getUserReplies: async (req, res, next) => {
-    const reqId = Number(req.params.id)
-    const reqUser = await User.findByPk(reqId)
-    if (!reqUser) return res.status(404).json({ status: 'error', message: 'User not found!' })
-    const replies = await Reply.findAll({
-      attributes: ['id', 'comment', 'createdAt'],
-      include: [
-        { model: User, attributes: ['id', 'name', 'account', 'avatar'] },
-        { model: Tweet, attributes: ['id'], include: { model: User, attributes: ['id', 'account'] } }
-      ],
-      where: { UserId: reqId },
-      order: [['createdAt', 'DESC']],
-      nest: true,
-      raw: true
-    })
-    const data = replies.map(reply => ({ ...reply, createdAt: dayjs(reply.createdAt).valueOf() }))
-    return res.status(200).json(data)
+    try {
+      const reqId = Number(req.params.id)
+      const reqUser = await User.findByPk(reqId)
+      if (!reqUser) return res.status(404).json({ status: 'error', message: 'User not found!' })
+
+      const replies = await Reply.findAll({
+        attributes: ['id', 'comment', 'createdAt'],
+        include: [
+          { model: User, attributes: ['id', 'name', 'account', 'avatar'] },
+          { model: Tweet, attributes: ['id'], include: { model: User, attributes: ['id', 'account'] } }
+        ],
+        where: { UserId: reqId },
+        order: [['createdAt', 'DESC']],
+        nest: true,
+        raw: true
+      })
+      const data = replies.map(reply => ({ ...reply, createdAt: dayjs(reply.createdAt).valueOf() }))
+
+      return res.status(200).json(data)
+    } catch (err) { next(err) }
   },
+
   getUserTweets: async (req, res, next) => {
     try {
       const reqId = Number(req.params.id)
       const loginUser = helpers.getUser(req)
       const reqUser = await User.findByPk(reqId)
       if (!reqUser) return res.status(404).json({ status: 'error', message: 'User not found!' })
+
       const tweets = await Tweet.findAll({
         attributes: ['id', 'description', 'createdAt',
           [sequelize.literal('(SELECT COUNT(*) FROM Replies WHERE Replies.TweetId = Tweet.id)'), 'replyCount'],
@@ -112,15 +107,18 @@ const userController = {
         isLiked: loginUser?.Likes?.some(loginUserLike => loginUserLike.TweetId === tweet.id),
         createdAt: dayjs(tweet.createdAt).valueOf()
       }))
+
       return res.status(200).json(data)
     } catch (err) { next(err) }
   },
+
   getUserLikes: async (req, res, next) => {
     try {
       const reqId = Number(req.params.id)
       const loginUser = helpers.getUser(req)
       const reqUser = await User.findByPk(reqId)
       if (!reqUser) return res.status(404).json({ status: 'error', message: 'User not found!' })
+
       const likes = await Like.findAll(
         {
           attributes: ['id', 'TweetId', 'createdAt'],
@@ -140,15 +138,18 @@ const userController = {
           raw: true
         }
       )
+
       const data = likes.map(like => {
         like.createdAt = dayjs(like.createdAt).valueOf()
         like.Tweet.createdAt = dayjs(like.Tweet.createdAt).valueOf()
         like.Tweet.isLiked = loginUser?.Likes?.some(loginUserLike => loginUserLike.TweetId === like.Tweet.id)
         return like
       })
+
       return res.status(200).json(data)
     } catch (err) { next(err) }
   },
+
   getUserFollowers: async (req, res, next) => {
     try {
       const reqId = Number(req.params.id)
@@ -174,6 +175,7 @@ const userController = {
       return res.status(200).json(data)
     } catch (err) { next(err) }
   },
+
   getUserFollowings: async (req, res, next) => {
     try {
       const reqId = Number(req.params.id)
@@ -188,34 +190,23 @@ const userController = {
         nest: true,
         raw: true
       })
+
       const data = followings.map(following => {
         following.createdAt = dayjs(following.createdAt).valueOf()
         following.updatedAt = dayjs(following.updatedAt).valueOf()
         following.Followings.isFollowed = loginUser?.Followings?.some(followingUser => followingUser.id === following.followingId)
         return following
       })
+
       return res.status(200).json(data)
     } catch (err) { next(err) }
   },
+
   putUserProfile: async (req, res, next) => {
     try {
       const profileId = Number(req.params.id)
-      const loginUserId = helpers.getUser(req).id
       const { name, introduction } = req.body
       const { files } = req
-
-      if (!name.trim()) {
-        return res.status(400).json({ status: 'error', message: 'User name is required!' })
-      }
-      if (name?.length > 50) {
-        return res.status(422).json({ status: 'error', message: 'Username should be less than 50 chars.' })
-      }
-      if (introduction?.length > 160) {
-        return res.status(422).json({ status: 'error', message: "User's introduction should be less than 160 chars" })
-      }
-      if (profileId !== loginUserId) {
-        return res.status(403).json({ status: 'error', message: "Cannot edit other user's profile." })
-      }
 
       const [user, avatarPath, coverPath] = await Promise.all([
         User.findByPk(profileId),
@@ -235,6 +226,7 @@ const userController = {
       return res.status(200).json({ status: 'success' })
     } catch (err) { next(err) }
   },
+
   getUsersTop: async (req, res, next) => {
     try {
       const loginUser = helpers.getUser(req)
@@ -246,13 +238,16 @@ const userController = {
         raw: true,
         nest: true
       })
+
       const data = topUsers.map(topuser => ({
         ...topuser,
         isFollowed: loginUser?.Followings?.some(followingUser => followingUser.id === topuser.id)
       }))
+
       return res.status(200).json(data)
     } catch (err) { next(err) }
   },
+
   getLoginUserProfile: (req, res, next) => {
     try {
       const loginUser = helpers.getUser(req).toJSON()
@@ -274,32 +269,12 @@ const userController = {
       return res.status(200).json(data)
     } catch (err) { next(err) }
   },
+
   putUserSetting: async (req, res, next) => {
     try {
-      const reqId = Number(req.params.id)
-      const loginUserId = helpers.getUser(req).id
-      const { account, name, email, password, checkPassword } = req.body
-      if (!account.trim() || !name.trim() || !email.trim() || !password.trim() || !checkPassword.trim()) {
-        return res.status(400).json({ status: 'error', message: 'All field are required!' })
-      }
-      if (!validator.isEmail(email)) {
-        return res.status(422).json({ status: 'error', message: 'Email input is invalid!' })
-      }
-      if (password !== checkPassword) {
-        return res.status(422).json({ status: 'error', message: 'Password and confirmPassword do not match.' })
-      }
-      if (name.length > 50) {
-        return res.status(422).json({ status: 'error', message: 'Name field has max length of 50 characters.' })
-      }
-      if (reqId !== loginUserId) {
-        return res.status(403).json({ status: 'error', message: "Cannot edit other user's setting." })
-      }
+      const { account, name, email, password } = req.body
 
-      const user = await User.findByPk(reqId)
-      if (!user) return res.status(404).json({ status: 'error', message: 'User not found!' })
-      if (user.email === email) return res.status(422).json({ status: 'error', message: 'Email already exists!' })
-      if (user.account === account) return res.status(422).json({ status: 'error', message: 'Account name already exists!' })
-
+      const user = req.userEdit
       const hash = await bcrypt.hash(password, 10)
       await user.update({ account, name, email, password: hash })
       return res.status(200).json({ status: 'success' })
